Clarify OpenAI SDK ownership and reasoning defaults
Sagent export branch: sagent/export/main
copybarista-source-rev-sha256=6aa8bacac8dae5d541280ad3ab521d1cb44298c3246cc842f41ba9bcd12e42c1
copybarista-replay-base-sha256=c684462ca9a54fb13706387b18da5674e09a74a19c97d4b21735ae6ad56f3892
copybarista-pr-state-version=1
