+++ sagent/providers/openai/api.py
@@ -1,81 +1,128 @@
-"""OpenAI provider.
-
-Usage::
-
-    from sagent.providers import OpenAI
-
-    provider = OpenAI.from_key("sk-...")
-    # or: export OPENAI_API_KEY=sk-... and use OpenAI.from_env()
-    gpt = provider.model("gpt-5.6-sol")
-    response = await gpt.buffer(request)
-"""
+"""OpenAI API-key provider using the Responses API."""
 
 from __future__ import annotations
 
 from collections.abc import Mapping
-from typing import ClassVar, Final
+from types import MappingProxyType
+from typing import TYPE_CHECKING, ClassVar, Self
+
+import os
 
 from sagent.catalog import openai as openai_catalog
-from sagent.providers.openai.compat import (
-    OpenAICompat,
-    OpenAICompatModel,
-)
+from sagent.providers.lib.perloop import PerLoop
+from sagent.providers.openai.responses import _OpenAIResponsesModel
 from sagent.types.capability import ModelCapability
+from sagent.types.providers import ModelRole, resolve
 
 
-__all__ = [
-    "OpenAI",
-]
+if TYPE_CHECKING:
+    import openai
+else:
+    from wrapt import lazy_import
+
+    openai = lazy_import("openai")
+
+__all__ = ["OpenAI"]
 
 
-class _OpenAIModel(OpenAICompatModel):
-    """OpenAI chat-completions backend - o-series accepts effort."""
+class OpenAI:
+    """API-key authentication and loop-local OpenAI SDK ownership."""
 
-
-# Conservative compatibility caps retained for pre-GPT-5.6 profiles:
-#   - ``max_image_dim`` = 2048: detail:high images are first scaled to fit
-#     within a 2048x2048 square before tiling
-#     (https://developers.openai.com/api/docs/guides/images-vision); pre-resizing
-#     to this caps wire bytes without losing fidelity the model would keep.
-#   - The 20 MB byte ceilings preserve their established local resize/compaction
-#     behavior. They are not the current GPT-5.6 limits; those are declared
-#     separately below.
-
-# GPT-5.6 ``auto``/``original`` preserves source dimensions. Current image
-# requirements publish a 512 MB TOTAL request-payload cap and no separate
-# per-image byte cap, so ``0`` keeps each image unmodified while the request
-# guard enforces the aggregate ceiling.
-# https://developers.openai.com/api/docs/guides/images-vision
-
-# Cheap-tier input window for models OpenAI prices in two tiers. For GPT-5.6,
-# gpt-5.5, gpt-5.4, and their applicable variants, prompts above 272K input
-# tokens bill at 2x input / 1.5x output for the whole session. The base profile
-# caps the window here; a ``+1m`` model id opts into the full window.
-# https://developers.openai.com/api/docs/models/gpt-5.6-sol
-_TWO_TIER_TOKENS: Final = 272_000
-
-
-class OpenAI(OpenAICompat):
-    """OpenAI provider."""
-
-    # Current frontier default for API-key users; ``+1m`` opts into the full
-    # 1.05M window. Subscription auth clamps this back to its 272K contract, so
-    # the suffix is a no-op on that path.
     DEFAULT_MODEL: ClassVar[str] = "gpt-5.6-sol+1m"
     DEFAULT_UTILITY_MODEL: ClassVar[str] = "gpt-5.4-mini"
-
     ENV_VAR: ClassVar[str] = "OPENAI_API_KEY"
     BASE_URL: ClassVar[str] = "https://api.openai.com/v1"
-
-    # Model limits and pricing.
-    # ModelLimits: https://developers.openai.com/api/docs/models/<model>
-    # Pricing: https://developers.openai.com/api/docs/pricing
-    # Cross-ref: https://github.com/taylorwilsdon/llm-context-limits
-    #
     CAPABILITIES: ClassVar[Mapping[str, ModelCapability]] = openai_catalog.models()
-    """Chat Completions rows: GPT-5.6 ``max`` downgrades to ``xhigh``."""
-
     TRANSPORT: ClassVar[ModelCapability] = openai_catalog.api()
-    """What this transport lets through; subclasses declare their own."""
 
-    MODEL_CLASS: ClassVar[type[OpenAICompatModel]] = _OpenAIModel
+    def __init__(self, *, api_key: str, base_url: str | None = None) -> None:
+        self.api_key = api_key
+        self.base_url = base_url or self.BASE_URL
+        self._sdks: PerLoop[openai.AsyncOpenAI] = PerLoop(self._make_sdk)
+
+    @property
+    def ROLES(self) -> Mapping[ModelRole, str]:  # noqa: N802 -- provider protocol spelling.
+        """Role names resolved by model construction."""
+        return MappingProxyType(
+            {"default": self.DEFAULT_MODEL, "utility": self.DEFAULT_UTILITY_MODEL}
+        )
+
+    @classmethod
+    def from_key(cls, api_key: str, *, base_url: str | None = None) -> Self:
+        """Construct a provider from an API key.
+
+        Args:
+          api_key: OpenAI API key.
+          base_url: Optional Responses endpoint override.
+
+        Returns:
+          provider: Configured provider.
+
+        """
+        return cls(api_key=api_key, base_url=base_url)
+
+    @classmethod
+    def from_env(cls, *, base_url: str | None = None) -> Self:
+        """Construct a provider from OPENAI_API_KEY.
+
+        Args:
+          base_url: Optional Responses endpoint override.
+
+        Returns:
+          provider: Configured provider.
+
+        """
+        key = os.environ.get(cls.ENV_VAR, "")
+        if not key:
+            raise RuntimeError(f"{cls.__name__} API key not configured.")
+        return cls(api_key=key, base_url=base_url)
+
+    def model(self, model_id: str | None = None) -> _OpenAIResponsesModel:
+        """Resolve a catalog model or role to a Responses backend.
+
+        Args:
+          model_id: Model id, context-tagged id, or role.
+
+        Returns:
+          model: Configured Responses model.
+
+        """
+        capability, settings = resolve(
+            model_id if model_id is not None else "default",
+            models=self.CAPABILITIES,
+            roles=self.ROLES,
+            transport=self.TRANSPORT,
+        )
+        return _OpenAIResponsesModel(
+            provider=self, capability=capability, settings=settings
+        )
+
+    def utility_model(self) -> _OpenAIResponsesModel:
+        """Return the utility-role model.
+
+        Returns:
+          model: Utility Responses backend.
+
+        """
+        return self.model("utility")
+
+    async def get_sdk(self) -> openai.AsyncOpenAI:
+        """Get this event loop's SDK client.
+
+        Returns:
+          sdk: Reusable client owned by this provider.
+
+        """
+        return self._sdks.get()
+
+    async def close_sdk(self) -> None:
+        """Close and release this loop's SDK without creating one."""
+        sdk = self._sdks.peek()
+        self._sdks.clear()
+        if sdk is not None:
+            await sdk.close()
+
+    def _make_sdk(self) -> openai.AsyncOpenAI:
+        return openai.AsyncOpenAI(
+            api_key=self.api_key, base_url=self.base_url, max_retries=0
+        )
