+++ sagent/providers/openai/sub.py
@@ -49,7 +49,7 @@ should prefer the API-key path (``OpenAI.from_key``).
 
 from __future__ import annotations
 
-from collections.abc import AsyncIterable, Callable, Mapping
+from collections.abc import Callable, Mapping
 from dataclasses import replace
 from pathlib import Path
 from types import MappingProxyType
@@ -59,7 +59,6 @@ from typing import (
     ClassVar,
     Final,
     NotRequired,
-    Protocol,
     TypedDict,
     cast,
     override,
@@ -69,7 +68,6 @@ from urllib.parse import quote_plus
 import asyncio
 import base64
 import contextlib
-import inspect
 import json
 import logging
 import os
@@ -79,42 +77,26 @@ import time
 
 
 if TYPE_CHECKING:
-    from openai.types.responses.response_input_param import FunctionCallOutput
+    from openai.types.responses.response_create_params import (
+        ResponseCreateParamsStreaming,
+    )
 
     import httpx2
     import openai
-    import openai.types.responses as oai_responses
-
-    import sagent.lib.image as image_lib
 else:
     from wrapt import lazy_import
 
     httpx2 = lazy_import("httpx2")  # 100ms cold
     openai = lazy_import("openai")  # 493ms cold
-    oai_responses = lazy_import("openai.types.responses")
-    image_lib = lazy_import("sagent.lib.image")
 
-from openai.lib.streaming.responses import AsyncResponseStream
 
 from sagent.catalog import openai as openai_catalog
-from sagent.lib import debug_log
 from sagent.lib.atomic_file import atomic_write_bytes
 from sagent.lib.custom_json import (
     DictCodec,
     FloatCodec,
-    IntCodec,
     MutableJSON,
-    json_unfreeze,
 )
-from sagent.providers.lib.errors import (
-    StreamingResponseNotReadError,
-    error_status_code,
-    find_response_not_read,
-    is_context_overflow_text,
-    is_request_too_large,
-    raise_if_request_too_large,
-)
-from sagent.providers.lib.id_remap import IdRemapper
 from sagent.providers.lib.oauth import (
     AuthCodeListener,
     credential_file_lock,
@@ -123,38 +105,24 @@ from sagent.providers.lib.oauth import (
     pkce_pair,
 )
 from sagent.providers.lib.perloop import PerLoop
-from sagent.providers.lib.stop_reason import normalize_stop_reason
-from sagent.providers.openai.api import OpenAI, _OpenAIModel
+from sagent.providers.openai.api import OpenAI
+from sagent.providers.openai.responses import _OpenAIResponsesModel
 from sagent.types.capability import (
     ContextTag,
     ModelCapability,
     ModelLimits,
 )
-from sagent.types.cost import TokenCount
 from sagent.types.exceptions import (
     AuthRefreshError,
-    UserFacingError,
 )
 from sagent.types.model import (
     ModelRequest,
     ModelResponse,
-    PromptTooLongError,
-    StreamInterruptedError,
-    base_model_id,
 )
 from sagent.types.providers import resolve
 from sagent.types.runtime import (
-    AgentSendMessage,
-    AssistantMessage,
-    BytesMessage,
-    ModelResponsePartial,
-    ModelResponseThinking,
     RuntimeEvent,
-    ToolCall,
-    ToolResult,
-    UserMessage,
 )
-from sagent.types.tools import Tool
 
 
 logger = logging.getLogger(__name__)
@@ -191,9 +159,6 @@ _SUBSCRIPTION_MAX_REQUEST_TOKENS = (
 _SUBSCRIPTION_MAX_RESPONSE_TOKENS = (
     32_000  # config-globals: ignore -- backend response budget, user-retunable
 )
-_STREAM_IDLE_TIMEOUT = (
-    600.0  # config-globals: ignore -- stream idle timeout, user-retunable
-)
 
 
 def _default_credentials_path() -> Path:
@@ -202,13 +167,6 @@ def _default_credentials_path() -> Path:
     if codex_home:
         return Path(codex_home).expanduser() / "auth.json"
     return DEFAULT_CREDENTIALS_PATH
-
-
-_FINISH_MAP: Final[dict[str, str]] = {
-    "completed": "stop",
-    "incomplete": "length",
-    "failed": "stop",
-}
 
 
 class _CredentialFileError(ValueError):
@@ -337,11 +295,8 @@ class OpenAISubscription(OpenAI):
 
     @classmethod
     @override
-    # The base returns `Self`; this deliberately returns a SIBLING (`OpenAI`),
-    # because subscription billing cannot use an API key. `Anthropic` avoids the
-    # warning only by redeclaring `from_key` with a concrete return, while
-    # `OpenAI` inherits `OpenAICompat`'s `Self` unchanged.
-    def from_key(  # ty: ignore[invalid-method-override] -- intentional sibling return; see above
+    # API-key authentication returns the base provider, not the OAuth subclass.
+    def from_key(  # ty: ignore[invalid-method-override] -- base returns Self; API keys cannot construct OAuth providers
         cls,
         api_key: str,
         *,
@@ -526,10 +481,9 @@ class OpenAISubscription(OpenAI):
         return creds
 
     @override
-    def model(  # ty: ignore[invalid-method-override] -- narrowed return type
+    def model(
         self,
         model_id: str | None = None,
-        /,
     ) -> _OpenAISubModel:
         """Create a Responses API model backend.
 
@@ -575,6 +529,7 @@ class OpenAISubscription(OpenAI):
         """True if the access token is within 5 min of expiry."""
         return time.time() > self._expires_at - self._refresh_buffer_sec
 
+    @override
     async def get_sdk(self) -> openai.AsyncOpenAI:
         """Return OAuth-authed SDK client, refreshing as needed.
 
@@ -609,6 +564,7 @@ class OpenAISubscription(OpenAI):
                 await cached[0].close()
             return sdk
 
+    @override
     async def close_sdk(self) -> None:
         """Close and clear the shared OAuth SDK client.
 
@@ -896,55 +852,17 @@ class OpenAISubscription(OpenAI):
         atomic_write_bytes(p, json.dumps(existing).encode(), file_mode=0o600)
 
 
-class _OpenAISubModel(_OpenAIModel):
-    """OpenAI Responses API model -- subscription billing."""
+class _OpenAISubModel(_OpenAIResponsesModel):
+    """OAuth retry and Codex-specific request restrictions."""
 
-    _provider: OpenAISubscription  # type: narrowed from OpenAICompat
-
-    @override
-    async def close(self) -> None:
-        """Release the provider's Responses-API SDK and the HTTP client.
-
-        The base ``OpenAICompatModel.close`` tears down the shared
-        ``httpx2.AsyncClient``; the subscription path additionally has an
-        ``AsyncOpenAI`` SDK owned by the *provider* (not the model), so
-        delegate to ``provider.close_sdk()`` for that.
-        """
-        await self._provider.close_sdk()
-        await super().close()
+    _provider: OpenAISubscription
 
     @override
-    def is_context_overflow(self, error: Exception) -> bool:
-        """Check whether an error indicates token context-window overflow.
-
-        Excludes the byte wire-limit via the shared classifier first, so a
-        413 byte error routes to byte-overflow recovery -- uniform with the
-        compat/google siblings -- rather than being mis-read as token
-        overflow because its body incidentally names the context window.
-
-        Args:
-          error: Exception raised by the API call.
-
-        Returns:
-          overflow: ``True`` if the error is a token context-length overflow.
-
-        """
-        if is_request_too_large(error_status_code(error), str(error)):
-            return False
-        return is_context_overflow_text(str(error))
-
-    @override
-    def is_retryable_provider_error(self, error: Exception) -> bool:
-        """OpenAI Responses API can return statusless transient errors.
-
-        The Codex subscription endpoint occasionally raises
-        ``openai.APIError`` with ``status_code=None`` and a body
-        containing the "An error occurred ... You can retry your
-        request" copy. The shared status-code classifier in
-        ``retry.py`` misses these because the status is absent;
-        match on the SDK's canonical retry-hint substring instead.
-        """
-        return "you can retry" in str(error).lower()
+    def _build_kwargs(self, request: ModelRequest) -> ResponseCreateParamsStreaming:
+        body = super()._build_kwargs(request)
+        body.pop("temperature", None)
+        body.pop("max_output_tokens", None)
+        return body
 
     @override
     async def stream(
@@ -952,677 +870,11 @@ class _OpenAISubModel(_OpenAIModel):
         request: ModelRequest,
         publish: Callable[[RuntimeEvent], None] | None = None,
     ) -> ModelResponse:
-        """Stream a request via the OpenAI Responses API.
-
-        Args:
-          request: Model request to send.
-          publish: Callback invoked with each runtime event as it arrives.
-
-        Returns:
-          response: Assembled model response after the stream closes.
-
-        """
-        sdk = await self._provider.get_sdk()
-        # The ChatGPT Codex subscription endpoint rejects some public
-        # Responses API knobs, including ``temperature`` and
-        # ``max_output_tokens``.
-        reasoning_effort = self._reasoning_effort()
-        # The Responses API only streams reasoning-text deltas when a
-        # ``summary`` is requested; without it the model reasons silently
-        # and ``on_thinking`` never fires. Request the auto summary whenever
-        # thinking is on so the reasoning surface is actually populated.
-        # (OpenAI never returns raw reasoning, only summaries; ``auto`` is
-        # the broadest tier and is chosen by the model.)
-        reasoning_dict: dict[str, str] | None = None
-        if reasoning_effort is not None:
-            reasoning_dict = {"effort": reasoning_effort, "summary": "auto"}
-            if base_model_id(self.capability.model_id).startswith(("gpt-5.6", "gpt-6")):
-                # Sagent replays complete local history with store=False, so
-                # these must render reasoning across all supplied turns.
-                # Verified accepted by gpt-6-astra on 2026-09-04.
-                reasoning_dict["context"] = "all_turns"
-        reasoning: dict[str, str] | openai.Omit = reasoning_dict or openai.omit
-        create_kwargs: dict[str, object] = {
-            "model": self._wire_model_id,
-            "input": _build_input(
-                request,
-                max_image_dim=self.limits.max_image_edge_px,
-                max_image_bytes=self.limits.max_image_bytes,
-            ),
-            "instructions": request.system or "",
-            "store": False,
-            "stream": True,
-            "include": ["reasoning.encrypted_content"]
-            if reasoning_effort is not None
-            else openai.omit,
-            "tools": _build_tools(request.tools) if request.tools else openai.omit,
-            "reasoning": reasoning,
-        }
-        tier = self._effective_service_tier()
-        if tier is not None:
-            create_kwargs["service_tier"] = tier
-        debug_log.trace(
-            "api_call",
-            kind="openai_responses",
-            model=self._wire_model_id,
-            service_tier=tier,
-        )
         try:
-            try:
-                event_stream: AsyncResponseStream = cast(
-                    AsyncResponseStream,
-                    await sdk.responses.create(**create_kwargs),  # pyright: ignore[reportCallIssue, reportArgumentType]  # ty: ignore[no-matching-overload] -- dynamic kwargs; SDK overload resolution fails on stream=True literal
-                )
-            except openai.AuthenticationError:
-                # In-memory bearer was revoked server-side between our
-                # local expiry check and the request arriving. Reload
-                # disk creds (or force-refresh) and retry once.
-                await self._provider.handle_auth_error()
-                sdk = await self._provider.get_sdk()
-                event_stream = cast(
-                    AsyncResponseStream,
-                    await sdk.responses.create(**create_kwargs),  # pyright: ignore[reportCallIssue, reportArgumentType]  # ty: ignore[no-matching-overload] -- dynamic kwargs; SDK overload resolution fails on stream=True literal
-                )
-            return await _consume_stream(
-                event_stream,
-                model=self,
-                publish=publish,
-            )
-        except Exception as exc:
-            not_read = find_response_not_read(exc)
-            if not_read is not None:
-                raise StreamingResponseNotReadError(
-                    provider_name="OpenAI subscription",
-                ) from exc
-            raise_if_request_too_large(error_status_code(exc), str(exc), cause=exc)
-            if self.is_context_overflow(exc):
-                # The compactor's shrink-and-retry path keys off
-                # PromptTooLongError; leaking raw SDK errors makes outer
-                # recovery retry unchanged history instead of dropping groups.
-                raise PromptTooLongError(str(exc)) from exc
-            raise
-
-    def _reasoning_effort(self) -> str | None:
-        """The wire reasoning effort, or ``None`` to omit reasoning."""
-        if self.settings.thinking_effort == "none":
-            return None
-        return openai_catalog.reasoning_effort(
-            self.settings.thinking_effort, model_id=self.capability.model_id
-        )
-
-
-def _build_tools(
-    tools: list[Tool],
-) -> list[oai_responses.FunctionToolParam]:
-    """Translate each ``Tool`` into a Responses API function-tool param."""
-    return [_build_tool(t) for t in tools]
-
-
-def _build_tool(tool: Tool) -> oai_responses.FunctionToolParam:
-    """Wire-shape a single ``Tool`` for the Responses API."""
-    return {
-        "type": "function",
-        "name": tool.name,
-        "description": tool.description,
-        "parameters": cast(
-            dict[str, object] | None, json_unfreeze(tool.directive_schema)
-        ),
-        "strict": None,
-    }
-
-
-def _build_input(
-    request: ModelRequest,
-    *,
-    max_image_dim: int = 0,
-    max_image_bytes: int = 0,
-) -> oai_responses.ResponseInputParam:
-    """Convert history entries to Responses API input items.
-
-    Args:
-      request: Fully-built model request.
-      max_image_dim: Maximum image dimension (pixels); larger inputs are
-          resized before encoding. ``0`` (the default) means no cap; the
-          live caller passes the model profile's ``max_image_dim``.
-      max_image_bytes: Maximum image size in bytes after resize. ``0`` (the
-          default) means no cap.
-
-    Returns:
-      items: Responses API input items in send order.
-
-    """
-    ids = IdRemapper("fc_")
-    items: oai_responses.ResponseInputParam = []
-    for entry in request.messages:
-        if isinstance(entry, (AgentSendMessage, UserMessage)):
-            items.append(
-                _build_user_item(entry, max_image_dim, max_image_bytes),
-            )
-        elif isinstance(entry, AssistantMessage):
-            _build_assistant_items(entry, items, ids)
-        else:
-            items.append(_build_tool_result_item(entry, ids))
-    return items
-
-
-def _build_user_item(
-    entry: AgentSendMessage | UserMessage,
-    max_image_dim: int,
-    max_image_bytes: int,
-) -> oai_responses.EasyInputMessageParam:
-    """Wire-shape a user-side entry, inlining any image attachments."""
-    image_atts = [att for att in entry.attachments if _is_image_attachment(att)]
-    non_image_atts = [att for att in entry.attachments if not _is_image_attachment(att)]
-    for att in non_image_atts:
-        logger.warning(
-            "OpenAI subscription: skipping non-image attachment (mime=%s)",
-            att.descriptor,
-        )
-    if not image_atts:
-        return {"role": "user", "content": entry.text}
-    blocks: list[oai_responses.ResponseInputContentParam] = []
-    if entry.text:
-        blocks.append({"type": "input_text", "text": entry.text})
-    for att in image_atts:
-        raw, mime = image_lib.resize(
-            att.data, max_dim=max_image_dim, max_bytes=max_image_bytes
-        )
-        b64 = base64.b64encode(raw).decode()
-        blocks.append(
-            {
-                "type": "input_image",
-                "detail": "auto",
-                "image_url": f"data:{mime};base64,{b64}",
-            }
-        )
-    return {"role": "user", "content": blocks}
-
-
-def _is_image_attachment(att: BytesMessage) -> bool:
-    """True for ``BytesMessage`` attachments whose descriptor is an image mime."""
-    return att.descriptor.startswith("image/")
-
-
-def _build_assistant_items(
-    entry: AssistantMessage,
-    items: oai_responses.ResponseInputParam,
-    ids: IdRemapper,
-) -> None:
-    """Expand an AssistantMessage into assistant + function_call items."""
-    for block in entry.thinking_blocks:
-        if block.get("type") == "reasoning" and isinstance(
-            block.get("encrypted_content"), str
-        ):
-            # OpenAI documents replaying the reasoning output item verbatim in
-            # stateless mode. Copy the opaque mapping so persisted session data
-            # never mutates while the SDK serializes the request.
-            items.append(cast(oai_responses.ResponseReasoningItemParam, dict(block)))
-    if entry.text:
-        items.append({"role": "assistant", "content": entry.text})
-    for tc in entry.tool_calls:
-        native_id = ids.map(tc.id)
-        items.append(
-            {
-                "type": "function_call",
-                "id": native_id,
-                "call_id": native_id,
-                "name": tc.name,
-                "arguments": json.dumps(dict(tc.args)),
-                "status": "completed",
-            }
-        )
-
-
-def _build_tool_result_item(
-    entry: ToolResult,
-    ids: IdRemapper,
-) -> FunctionCallOutput:
-    """Wire-shape a ``ToolResult`` as a Responses API function_call_output."""
-    native_id = ids.map(entry.call_id)
-    output = entry.content
-    if entry.is_error and output:
-        output = f"[Error] {output}"
-    return {
-        "type": "function_call_output",
-        "call_id": native_id,
-        "output": output,
-        "status": "completed",
-    }
-
-
-class _UsageDetails(Protocol):
-    """Minimum SDK usage-details surface needed for forward-compatible fields."""
-
-    def to_dict(self) -> dict[str, object]: ...
-
-
-def _terminal_metadata(response: object) -> tuple[str, str, int, int, int, int]:
-    """Extract terminal response metadata, including SDK-extra usage fields."""
-    message_id = str(getattr(response, "id", "") or "")
-    status = str(getattr(response, "status", "") or "")
-    usage = getattr(response, "usage", None)
-    if usage is None:
-        return message_id, status, 0, 0, 0, 0
-    input_tokens = IntCodec.coerce(getattr(usage, "input_tokens", None), 0)
-    output_tokens = IntCodec.coerce(getattr(usage, "output_tokens", None), 0)
-    cache_read = 0
-    cache_write = 0
-    details = getattr(usage, "input_tokens_details", None)
-    if details is not None:
-        # OpenAI SDK 2.44 predates the typed ``cache_write_tokens`` field, but
-        # Stainless models preserve unknown response fields and expose them
-        # through ``to_dict``. Reading the mapping keeps new usage data without
-        # an Any cast or waiting for an SDK schema release.
-        raw_details = cast(_UsageDetails, details).to_dict()
-        cache_read = IntCodec.coerce(raw_details.get("cached_tokens"), 0)
-        cache_write = IntCodec.coerce(raw_details.get("cache_write_tokens"), 0)
-    return (
-        message_id,
-        status,
-        input_tokens,
-        output_tokens,
-        cache_read,
-        cache_write,
-    )
-
-
-_STREAM_ERROR_STATUS: Final[dict[str, int]] = {
-    "rate_limit": 429,
-    "rate_limit_error": 429,
-    "rate_limit_exceeded": 429,
-    "api_error": 500,
-    "overloaded_error": 529,
-    "server_error": 500,
-}
-
-
-class _OpenAIStreamError(UserFacingError):
-    """In-band Responses error retaining retry/rate-limit classification data."""
-
-    def __init__(
-        self,
-        message: str,
-        *,
-        code: str | None,
-        param: str | None = None,
-    ) -> None:
-        super().__init__(message)
-        normalized_code = code or "unknown_error"
-        normalized_type = (
-            "rate_limit_error"
-            if normalized_code in {"rate_limit", "rate_limit_exceeded"}
-            else normalized_code
-        )
-        self.status_code = _STREAM_ERROR_STATUS.get(normalized_code)
-        self.body: Mapping[str, object] = {
-            "type": "error",
-            "error": {
-                "type": normalized_type,
-                "code": normalized_code,
-                "message": message,
-                "param": param,
-            },
-        }
-
-
-async def _consume_stream(
-    stream: AsyncIterable[object],
-    *,
-    model: _OpenAISubModel,
-    publish: Callable[[RuntimeEvent], None] | None,
-) -> ModelResponse:
-    """Parse a Responses API event stream into an assembled ModelResponse."""
-    text_parts: list[str] = []
-    refusal_parts: list[str] = []
-    thinking_parts: list[str] = []
-    encrypted_reasoning: list[Mapping[str, object]] = []
-    tool_calls: list[ToolCall] = []
-    tool_args: dict[str, list[str]] = {}
-    input_tokens = 0
-    output_tokens = 0
-    cache_read = 0
-    cache_write = 0
-    message_id = ""
-    finish_reason: str | None = None
-    completed = False
-
-    loop = asyncio.get_running_loop()
-    deadline = loop.time() + _STREAM_IDLE_TIMEOUT
-    # Single cleanup rule: every path that does not return a terminal response
-    # closes the stream. A mid-stream error event (raises below), an idle
-    # timeout / cancellation, and the no-completion fallthrough all leave the
-    # SSE connection open otherwise -- a per-error connection leak. ``completed``
-    # gates the one exit that legitimately keeps reading until done.
-    try:
-        async with asyncio.timeout_at(deadline) as watchdog:
-            async for event in stream:
-                watchdog.reschedule(loop.time() + _STREAM_IDLE_TIMEOUT)
-                if isinstance(event, oai_responses.ResponseRefusalDeltaEvent):
-                    refusal_parts.append(event.delta)
-                    text_parts.append(event.delta)
-                    if publish is not None:
-                        publish(ModelResponsePartial(event.delta))
-                elif isinstance(event, oai_responses.ResponseRefusalDoneEvent):
-                    # Normally deltas already contain the full refusal. Keep a
-                    # done-only stream useful and append only an unseen suffix.
-                    seen = "".join(refusal_parts)
-                    suffix = (
-                        event.refusal[len(seen) :]
-                        if event.refusal.startswith(seen)
-                        else ""
-                    )
-                    if not seen:
-                        suffix = event.refusal
-                    if suffix:
-                        refusal_parts.append(suffix)
-                        text_parts.append(suffix)
-                        if publish is not None:
-                            publish(ModelResponsePartial(suffix))
-                elif isinstance(event, oai_responses.ResponseTextDeltaEvent):
-                    text_parts.append(event.delta)
-                    if publish is not None:
-                        publish(ModelResponsePartial(event.delta))
-                elif isinstance(
-                    event,
-                    (
-                        oai_responses.ResponseReasoningTextDeltaEvent,
-                        oai_responses.ResponseReasoningSummaryTextDeltaEvent,
-                    ),
-                ):
-                    thinking_parts.append(event.delta)
-                    if publish is not None:
-                        publish(ModelResponseThinking(event.delta))
-                elif isinstance(
-                    event,
-                    oai_responses.ResponseFunctionCallArgumentsDeltaEvent,
-                ):
-                    tool_args.setdefault(event.item_id, []).append(event.delta)
-                elif isinstance(event, oai_responses.ResponseOutputItemDoneEvent):
-                    item = event.item
-                    if item.type == "function_call":
-                        tc_id = str(item.call_id or "")
-                        tc_name = str(item.name or "")
-                        item_id = item.id or ""
-                        delta_args = "".join(tool_args.get(item_id, []))
-                        done_args = str(item.arguments or "")
-                        args = _parse_tool_arguments(
-                            delta_args,
-                            done_args,
-                            tool_name=tc_name,
-                            call_id=tc_id,
-                        )
-                        tool_calls.append(
-                            ToolCall(
-                                id=tc_id,
-                                name=tc_name,
-                                args=cast(Mapping[str, object], args),
-                            )
-                        )
-                    elif item.type == "reasoning":
-                        reasoning_item = item.to_dict(exclude_none=True)
-                        if isinstance(reasoning_item.get("encrypted_content"), str):
-                            encrypted_reasoning.append(reasoning_item)
-                elif isinstance(event, oai_responses.ResponseErrorEvent):
-                    raise _openai_stream_event_error(event)
-                elif isinstance(event, oai_responses.ResponseFailedEvent):
-                    raise _openai_stream_response_error(event.response)
-                elif isinstance(
-                    event,
-                    (
-                        oai_responses.ResponseCompletedEvent,
-                        oai_responses.ResponseIncompleteEvent,
-                    ),
-                ):
-                    resp = event.response
-                    (
-                        message_id,
-                        finish_reason,
-                        input_tokens,
-                        output_tokens,
-                        cache_read,
-                        cache_write,
-                    ) = _terminal_metadata(resp)
-                    if finish_reason == "incomplete":
-                        incomplete = getattr(resp, "incomplete_details", None)
-                        if getattr(incomplete, "reason", None) == "content_filter":
-                            finish_reason = "content_filter"
-                    debug_log.trace(
-                        "api_response",
-                        kind="openai_responses",
-                        service_tier=getattr(resp, "service_tier", None),
-                    )
-                    # Set last: if a usage/attr read above raises on an SDK shape
-                    # change, ``completed`` stays False so the ``finally`` still
-                    # closes the stream instead of leaking the connection.
-                    completed = True
-    finally:
-        if not completed:
-            await _close_stream(stream)
-
-    if refusal_parts and finish_reason in (None, "completed", "stop"):
-        finish_reason = "content_filter"
-    response = _build_stream_response(
-        text_parts=text_parts,
-        thinking_parts=thinking_parts,
-        encrypted_reasoning=encrypted_reasoning,
-        tool_calls=tool_calls,
-        input_tokens=input_tokens,
-        output_tokens=output_tokens,
-        cache_write=cache_write,
-        cache_read=cache_read,
-        finish_reason=finish_reason,
-        message_id=message_id,
-        model=model,
-    )
-    if not completed:
-        raise StreamInterruptedError(response)
-    return response
-
-
-def _openai_stream_event_error(event: object) -> UserFacingError:
-    code = getattr(event, "code", None)
-    param = getattr(event, "param", None)
-    message = getattr(event, "message", "OpenAI subscription stream error")
-    details = ["OpenAI subscription stream error"]
-    if isinstance(code, str) and code:
-        details.append(f"code={code}")
-    if isinstance(param, str) and param:
-        details.append(f"param={param}")
-    details.append(str(message))
-    return _OpenAIStreamError(
-        ": ".join(details),
-        code=code if isinstance(code, str) else None,
-        param=param if isinstance(param, str) else None,
-    )
-
-
-def _openai_stream_response_error(response: object) -> UserFacingError:
-    response_id = getattr(response, "id", "")
-    status = getattr(response, "status", "")
-    error = getattr(response, "error", None)
-    message = getattr(error, "message", None) if error is not None else None
-    code = getattr(error, "code", None) if error is not None else None
-    incomplete = getattr(response, "incomplete_details", None)
-    reason = getattr(incomplete, "reason", None) if incomplete is not None else None
-    details = ["OpenAI subscription stream ended without completion"]
-    if isinstance(status, str) and status:
-        details.append(f"status={status}")
-    if isinstance(response_id, str) and response_id:
-        details.append(f"response_id={response_id}")
-    if isinstance(code, str) and code:
-        details.append(f"code={code}")
-    if isinstance(reason, str) and reason:
-        details.append(f"reason={reason}")
-    if isinstance(message, str) and message:
-        details.append(message)
-    return _OpenAIStreamError(
-        ": ".join(details),
-        code=code if isinstance(code, str) else None,
-    )
-
-
-def _build_stream_response(
-    *,
-    text_parts: list[str],
-    thinking_parts: list[str],
-    encrypted_reasoning: list[Mapping[str, object]],
-    tool_calls: list[ToolCall],
-    input_tokens: int,
-    output_tokens: int,
-    cache_write: int,
-    cache_read: int,
-    finish_reason: str | None,
-    message_id: str,
-    model: _OpenAISubModel,
-) -> ModelResponse:
-    raw_reason = _FINISH_MAP.get(finish_reason or "", finish_reason)
-
-    # The Responses API reports a cache-inclusive prompt total; keep ordinary,
-    # cache-write, and cache-read pools disjoint in Sagent's normalized usage.
-    non_cached_input = max(0, input_tokens - cache_read - cache_write)
-    tokens = TokenCount(
-        request=non_cached_input,
-        response=output_tokens,
-        cache_write=cache_write,
-        cache_read=cache_read,
-    )
-    thinking_blocks: list[Mapping[str, object]] = list(encrypted_reasoning)
-    if thinking_parts:
-        thinking_blocks.append({"type": "reasoning", "text": "".join(thinking_parts)})
-    return ModelResponse(
-        message=AssistantMessage(
-            text="".join(text_parts),
-            thinking_blocks=tuple(thinking_blocks),
-            tool_calls=tuple(tool_calls),
-        ),
-        tokens=tokens,
-        stop_reason=normalize_stop_reason(
-            raw_reason,
-            kind="openai",
-            has_tool_use=bool(tool_calls),
-        ),
-        message_id=message_id,
-        request_id=message_id,
-        spend=model.spend(tokens),
-    )
-
-
-async def _close_stream(stream: object) -> None:
-    """Close a Responses stream after timeout, cancellation, or stream error.
-
-    Runs in a ``finally`` / cleanup context, so a failure HERE must never
-    replace the error that triggered the close (e.g. a ``UserFacingError`` from a
-    ``ResponseFailedEvent``). A broken pipe on ``aclose`` would otherwise surface
-    to the user instead of the real cause -- swallow and log it.
-    """
-    close = getattr(stream, "aclose", None)
-    if close is None:
-        close = getattr(stream, "close", None)
-    if close is None:
-        return
-    try:
-        result = close()
-        if inspect.isawaitable(result):
-            await result
-    except Exception:
-        logger.debug("stream close raised during cleanup", exc_info=True)
-
-
-def _parse_tool_arguments(
-    delta_args: str,
-    done_args: str,
-    *,
-    tool_name: str,
-    call_id: str,
-) -> MutableJSON:
-    """Parse streamed function-call arguments with completed-item fallback."""
-    saw_args = False
-    parsed_by_source: dict[str, MutableJSON] = {}
-    for source, args_str in (("delta", delta_args), ("done", done_args)):
-        if not args_str:
-            continue
-        saw_args = True
-        try:
-            parsed = json.loads(args_str)
-        except json.JSONDecodeError:
-            logger.warning(
-                "OpenAI Responses tool arguments were invalid JSON: "
-                "source=%s tool=%s call_id=%s chars=%d",
-                source,
-                tool_name,
-                call_id,
-                len(args_str),
-            )
-            continue
-        if isinstance(parsed, dict):
-            parsed_obj = cast(MutableJSON, parsed)
-            parsed_by_source[source] = parsed_obj
-            if not parsed_obj:
-                logger.warning(
-                    "OpenAI Responses tool arguments were an empty JSON object: "
-                    "source=%s tool=%s call_id=%s chars=%d",
-                    source,
-                    tool_name,
-                    call_id,
-                    len(args_str),
-                )
-            continue
-        logger.warning(
-            "OpenAI Responses tool arguments were not a JSON object: "
-            "source=%s tool=%s call_id=%s",
-            source,
-            tool_name,
-            call_id,
-        )
-    if not saw_args:
-        logger.warning(
-            "OpenAI Responses tool arguments were empty: tool=%s call_id=%s",
-            tool_name,
-            call_id,
-        )
-    delta = parsed_by_source.get("delta")
-    done = parsed_by_source.get("done")
-    if delta is not None and done is not None and delta != done:
-        logger.warning(
-            "OpenAI Responses tool arguments differed between delta and done: "
-            "tool=%s call_id=%s delta_chars=%d done_chars=%d "
-            "delta_keys=%d done_keys=%d",
-            tool_name,
-            call_id,
-            len(delta_args),
-            len(done_args),
-            len(delta),
-            len(done),
-        )
-    if done:
-        logger.debug(
-            "OpenAI Responses selected completed tool arguments: "
-            "tool=%s call_id=%s delta_chars=%d done_chars=%d "
-            "delta_keys=%d done_keys=%d",
-            tool_name,
-            call_id,
-            len(delta_args),
-            len(done_args),
-            len(delta) if delta is not None else -1,
-            len(done),
-        )
-        return done
-    if delta:
-        logger.debug(
-            "OpenAI Responses selected streamed delta tool arguments: "
-            "tool=%s call_id=%s delta_chars=%d done_chars=%d delta_keys=%d",
-            tool_name,
-            call_id,
-            len(delta_args),
-            len(done_args),
-            len(delta),
-        )
-        return delta
-    if done is not None:
-        return done
-    if delta is not None:
-        return delta
-    return {}
+            return await super().stream(request, publish)
+        except openai.AuthenticationError:
+            await self._provider.handle_auth_error()
+        return await super().stream(request, publish)
 
 
 def _jwt_payload(token: str) -> dict[str, object]:
