+++ sagent/providers/wire_conformance_test.py
@@ -34,6 +34,7 @@ from sagent.providers.minimax.api import MiniMax
 from sagent.providers.moonshot.api import Moonshot
 from sagent.providers.openai.api import OpenAI
 from sagent.providers.openai.compat import OpenAICompatModel
+from sagent.providers.openai.responses import _OpenAIResponsesModel
 from sagent.types.capability import (
     ContextTag,
     ModelCapability,
@@ -95,6 +96,12 @@ def _google_thinking(model: Model, settings: ModelSettings) -> object:
     return gen_config.get("thinkingConfig")
 
 
+def _responses_thinking(model: Model, settings: ModelSettings) -> object:
+    assert isinstance(model, _OpenAIResponsesModel)
+    model._settings = settings
+    return model._build_kwargs(_request()).get("reasoning")
+
+
 def _chat_thinking(model: Model, settings: ModelSettings) -> object:
     """Chat-completions reasoning knobs, or ``None`` when it sends none."""
     assert isinstance(model, OpenAICompatModel)
@@ -117,7 +124,7 @@ _WireBuilder = tuple[
 _WIRE_BUILDERS: Mapping[str, _WireBuilder] = {
     "Anthropic": (lambda: Anthropic.from_key("k"), _anthropic_thinking),
     "Google": (lambda: Google.from_key("k"), _google_thinking),
-    "OpenAI": (lambda: OpenAI.from_key("k"), _chat_thinking),
+    "OpenAI": (lambda: OpenAI.from_key("k"), _responses_thinking),
     "DashScope": (lambda: DashScope.from_key("k"), _chat_thinking),
     "MiniMax": (lambda: MiniMax.from_key("k"), _chat_thinking),
     "Moonshot": (lambda: Moonshot.from_key("k"), _chat_thinking),
